FuryAnnotationTest
