+++ java/fury-core/src/test/java/org/apache/fury/FuryTest.java
@@ -27,7 +27,6 @@ import static org.testng.Assert.assertTrue;
 import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Lists;
 import java.io.*;
 import java.lang.invoke.MethodHandles;
 import java.math.BigDecimal;
@@ -54,7 +53,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import org.apache.fury.annotation.Expose;
-import org.apache.fury.annotation.FuryField;
 import org.apache.fury.annotation.Ignore;
 import org.apache.fury.builder.Generated;
 import org.apache.fury.config.CompatibleMode;
@@ -355,148 +353,6 @@ public class FuryTest extends FuryTestBase {
     o.f1 = 10;
     o.f2 = 1;
     serDeCheckSerializer(fury, o, "PackageLevelBean");
-  }
-
-  @Data
-  public static class BeanM {
-    @FuryField(nullable = false)
-    public Long f1;
-
-    @FuryField(nullable = false)
-    private Long f2;
-
-    String s = "str";
-    Short shortValue = Short.valueOf((short) 2);
-    Byte byteValue = Byte.valueOf((byte) 3);
-    Long longValue = Long.valueOf(4L);
-    Boolean booleanValue = Boolean.TRUE;
-    Float floatValue = Float.valueOf(5.0f);
-    Double doubleValue = Double.valueOf(6.0);
-    Character character = Character.valueOf('c');
-
-    int i = 10;
-
-    int i2;
-
-    long l1;
-
-    double d1;
-
-    char c1;
-
-    boolean b1;
-
-    byte byte1;
-
-    @FuryField(nullable = false)
-    List<Integer> integerList = Lists.newArrayList(1);
-
-    @FuryField(nullable = false)
-    String s1 = "str";
-
-    @FuryField(nullable = false)
-    Short shortValue1 = Short.valueOf((short) 2);
-
-    @FuryField(nullable = false)
-    Byte byteValue1 = Byte.valueOf((byte) 3);
-
-    @FuryField(nullable = false)
-    Long longValue1 = Long.valueOf(4L);
-
-    @FuryField(nullable = false)
-    Boolean booleanValue1 = Boolean.TRUE;
-
-    @FuryField(nullable = false)
-    Float floatValue1 = Float.valueOf(5.0f);
-
-    @FuryField(nullable = false)
-    Double doubleValue1 = Double.valueOf(6.0);
-
-    @FuryField(nullable = false)
-    Character character1 = Character.valueOf('c');
-
-    @FuryField(nullable = true)
-    List<Integer> integerList1 = Lists.newArrayList(1);
-
-    @FuryField(nullable = true)
-    String s2 = "str";
-
-    @FuryField(nullable = true)
-    Short shortValue2 = Short.valueOf((short) 2);
-
-    @FuryField(nullable = true)
-    Byte byteValue2 = Byte.valueOf((byte) 3);
-
-    @FuryField(nullable = true)
-    Long longValue2 = Long.valueOf(4L);
-
-    @FuryField(nullable = true)
-    Boolean booleanValue2 = Boolean.TRUE;
-
-    @FuryField(nullable = true)
-    Float floatValue2 = Float.valueOf(5.0f);
-
-    @FuryField(nullable = true)
-    Double doubleValue2 = Double.valueOf(6.0);
-
-    @FuryField(nullable = true)
-    Character character2 = Character.valueOf('c');
-
-    public BeanM() {
-      this.f1 = 1L;
-      this.f2 = 1L;
-    }
-  }
-
-  @Data
-  public static class BeanN {
-    public long f1;
-    private long f2;
-  }
-
-  @Data
-  public static class BeanM1 {
-
-    @FuryField(nullable = false)
-    private BeanN beanN;
-  }
-
-  @Test(dataProvider = "basicMultiConfigFury")
-  public void testFuryFieldAnnotation(
-      boolean trackingRef,
-      boolean codeGen,
-      boolean scopedMetaShare,
-      CompatibleMode compatibleMode) {
-    Fury fury =
-        Fury.builder()
-            .withLanguage(Language.JAVA)
-            .withRefTracking(trackingRef)
-            .requireClassRegistration(false)
-            .withCodegen(codeGen)
-            .withCompatibleMode(compatibleMode)
-            .withScopedMetaShare(scopedMetaShare)
-            .build();
-    BeanM o = new BeanM();
-    byte[] bytes = fury.serialize(o);
-    final Object deserialize = fury.deserialize(bytes);
-    Assert.assertEquals(o, deserialize);
-  }
-
-  @Test(dataProvider = "referenceTrackingConfig")
-  public void testFuryFieldAnnotationException(boolean referenceTracking) {
-    Fury fury =
-        Fury.builder()
-            .withLanguage(Language.JAVA)
-            .withRefTracking(referenceTracking)
-            .requireClassRegistration(false)
-            .withCodegen(false)
-            .build();
-    BeanM1 o1 = new BeanM1();
-    if (referenceTracking) {
-      assertEquals(serDe(fury, o1), o1);
-    } else {
-      assertThrows(NullPointerException.class, () -> fury.serialize(o1));
-    }
   }
 
   static class B {
